+++ src/model/entities/Princess.java
@@ -4,7 +4,7 @@ import java.awt.Dimension;
 
 public class Princess extends EntityImpl{
 
-    public Princess(final int x,final int y, final Dimension dim) {
+    public Princess(final Double x,final Double y, final Dimension dim) {
         super(x, y, dim);
     }
